# Backend_v2/openApi.yaml
...
openapi: '3.0.3'
info:
  title: Bloglite API
  version: '1.0'
servers:
  - url: http://127.0.0.1:5000
    description: Local server
security:
  - bearerAuth: []
...
paths:
...
  /bloglite/user:
    get:
      tags:
        - User
      summary: User data
      description: get all the user data
      operationId: user Data
      security:
        - bearerAuth: []
      responses: 
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found"
  /bloglite/registration:
    post:
      tags:
        - User
      summary: Create user
      description: Create a new user
      operationId: new user
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Random Person
                username:
                  type: string
                  example: user123
                password:
                  type: string
                  format: password
                  example: password@123
      responses:
        '201':
          description:
            Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User Created successfully"
        
        '400':
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "user already exists"
  /bloglite/password_update:
    put:
      tags:
        - User
      summary: Update password
      description: Create a new password
      operationId: new password

                
      security:
        - bearerAuth: []
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  format: password
                  example: Password@123
                new_password:
                  type: string
                  format: password
                  example: Password@123
      responses:
        '200':
          description:
            Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password Updated Successfully"
        
        '400':
          description: Inavlid Password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid Password"
  /bloglite/deleteuser:
    delete:
      tags:
        - User
      summary: Delete your profile
      description: It will erase your profile
      operationId: Delete user

                
      security:
        - bearerAuth: []

      responses:
        '200':
          description:
            Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted successfully"
        
        '400':
          description: Something went wrong
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Something went worng"


# Feeds
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas: 
    User:
      type: object
      properties: 
        userName:
          type: string
          example: Random user
        userUsername:
          type: string
          example: ram123
        userNoBlog:
          type: integer
          format: int64
          example: 20
        userNoFollowers:
          type: integer
          format: int64
          example: 56
        userNoFollowing:
          type: integer
          format: int64
          example: 44
...